refactor: extract project form schemas into dedicated modules
Move the Zod schemas and body-length constants out of ProjectUpdate.tsx
and ProjectDialog/index.tsx into sibling schema modules. This keeps the
component files exporting only components (fixes the React Fast Refresh
only-export-components error) and brings both files back under the
file-size baseline. Schema-only tests move to a focused test module.
